VIX-2226 Restructure how the curve and gradient edit works.
There is no need to iterate the library to find the curve edited by name when it has a method to directly retrieve it by name. This should make edits of large libraries more efficient.

Remove the logic that was calling end bulk updates on the libraries in the PopulateToolStripClass method. There were not any updates occurring in that method and it was causing unwanted behaviors with the new event scheme.