VIX-2741 Correct the issues that caused Vixen to crash when expected folders did not exist.
The config folder for the Falcon universe export was expected to exist, but may not if the user exports to a non Falcon folder structure. Change the logic to create paths where necessary to ensure success. The top level path is prompted if it does not exist in the wizard, and then sub paths are auto created.