VIX-2234 Add the UI bits in the web interface to drive the controller API.
Finished up the UI parts to drive the new Web API functions for the controllers. Support is enabled for individual controllers to be turned on and off as well as all of them at the same time. The section is collapsible to keep it out of the way when not needed.