VIX-2269 Numerous fixes to improve the performance of operations in the Display Setup.
Reduce the places that fully reload the controller tree or the elements tree. Create methods to refresh the portions of the trees that are effected by changes. Add some additional event handling arguments to the elements changed to better describe how it changed so the logic can better react. Clean up some extraneous events that caused logic to run numerous times when it was not necessary.