VIX-2777  Initial work on a vendor model browser.
This currently provides a way to read a xml representation of the vendors inventory along with being able to download and import a xModel into the custom prop editor to transform it into a Vixen prop model. The UI can be used in the future for our own format as the proprietary xml format used now is translated into a more universal notion to abstract it away.

Adjust the origin of the imported model offset a small amount from 0,0 to make it easier to move it around.