VIX-2759  Initial work to tweak on some performance aspects to squeeze out a bit more life.
Probably should have broken this commit up, but there is a good amount of cross commonality in this.

Remove a bunch of dead code classes to reduce the bloat of the project. Remove some unused methods from the intent objects to simplify them.

Optimized the LightingValue to use a simple color instead of converting to HSV which required more conversions elsewhere and is a larger payload. Should cut down on some memory use. Step one in moving to a unified color structure. Optimized the logic in the LightingValueInterpolator with the changes to the color structure in the LightingValue, the interpolation logic can simplify. A similar change was made to the ColorBreakdown. More changes to come in that object.

Fixed up the static intent so it can be used in the set level instead of the linear intent which has unnecessary over head. Made an optimization in the intents to ensure they are only interpolated once per time offset. Before each consumer would incur the interpolation cost. Although not overly expensive, it could add up on larger displays.

Optimized the output filter in the breakdown. When mixing colors, the logic could be greatly simplified. Made an additional filter specifically for mixing that does not have all the branch logic in it to see if we are doing non mixing types which are the minority.