VIX-2759 Removed some additional dead code projects.
Removed the Direct2D project and related dependencies as they are not being used anywhere.

Added clean up steps to the build of the preview, editor, and custom prop editor to clean up a bunch of dependencies it copies locally when it should not be.

Removed a NLog reference from the ChromaKey project that was incorrect and not being used.