VIX-2651 Work to try and close the gap between sequences in the scheduler.
Work on part of the reason that there was a cycle delay in the scheduler by making sure that the events in the executor fire at the correct time.
Remove the delay timer in the scheduler action that was trying to wait out the issue above to close the gap. This brings it very close to transition most of the time that is virtually seamless.