VIX-2731 Effect Editor changes to better manage collections, and style improvements.
Make some changes to how the Collection types are displayed in the editor. In the past they would have a single property name on the left and all the values would be bunched to the right. This will move the label from the left to a overall header and allow the collection items to take up the full space and be first class citizens.

The new Emitter collection editor gains a numbered header rule for each Emitter in the collection to help separate them out for the user. There are a lot of properties in each one and it could be a bit confusing to determine where they start and end. This could be extended to other collection types.

The Emitter type gets some styling adjustments to bring it in line with the remainder of the property names. Swap out the TextBlock and styles for the already defined PropertyNameTextBlock that has the style elements of the remainder of the labels. This provides a nice consistency.

Update gitignore to ignore the output folders from the C++ projects. Feels like these are outputting to weird places, but will address that later.

Remove the copy local attribute from the color gradient dependency in the Liquid project to prevent module loader breakage.