VIX-2776 Clean up build dependencies.
Ensure all the targets for the QM Library project are building against v141 tools.

Removed several unused dependencies. Old Git stuff, DirectX deps from old preview. Code tools from when we allowed scripting in an effect.

Removed build output clean up steps now that we are on 4.7.2 which has more support for .Net Core 2.0.

Removed the mrdsio/inout libraries for the RDS calls that are not being used. Commented the code in case a need arises to put it back. It was calling disconnect still, but I don't think that was necessary as the serial port logic closes the connection.