VIX-2731 Initial work to add functionality to the editor to better handle complex objects in collections.
Start work on enhancing the CollectionItemValue class to discover complex object properties. This may get refactored to another location depending on how it works out, but this will see if it is viable.

Refactor the Emitter editor to use the property Item layout so the automatic stuff will work. This likely needs more work to finish it out.

Add some of the Attributes to the Emitter class to test some of the other abilities they provide for.

TODO manage value changed on SubProperties and propagate those up the tree so things outside the editor can be aware they were updated.