VIX-2731 Add logic for more feature rich collection editing.
Refactor the CollectionItemValue to manage objects marked with the ExpandoObject attribute as deep properties. Interrogate and load them up as regular property items so all the magic works. Add logic to manage the edit events and propagate them up. Add logic to listen for the TypeDescriptor refresh.

Add a base expando object that provides for type descriptor handling and property notification.

Refactor the logic in the Emitter to use the attribute based mark ups for name and description as well as editor preference. Utilize the new base expando object for the type descriptor goodness and the property changed support. This surly needs more work, but is a good start.