VIX-2731 Debounce the selection change event in the editor.
The Effect Editor can get multiple selection changed events in rapid fire and this causes extra work. We really only need to respond the the last in a string of them. This relieves some of the pressure on editing the new Liquid effect that has a much more complex property tree. It will have broad benefit to others as well even though they are less complex and generally faster.