VIX-2731 Additional usability items.
Add in some logic to forward the mouse wheel events up the tree in the collection views. When you are over the larger list items now, it was hard to find an area that he scroll wheel would work off to the sides. This allows scrolling over the items and the whole pane scrolls. I extended this to the other collection view editors as well.

Change the list item hover style to be just a simple outline instead of the default shaded background. We turned this off on the other ones, but this one seemed like a simple outline would be beneficial to see where the larger set of the Emitter started and ended. Along with the header this makes it a little more distinct. I turned the same on for the gradient level pairs as they could benefit from the same thing. I left it off the single items types.