VIX-2778 Fix issues in the Custom Prop Editor with xModel imports.
Handle closing tags properly on Submodel entries. Sometimes they are inline as one tag and other times they have children. The parser was accepting the closing tag and trying to get attributes from them which ended up leading to a crash.

Ranges were assumed to be increment in nature, but some models have ranges that decrement. Add logic to handle the decrement cases properly.

Fix the binding error that was not allowing the Order number to appear in the order list. Fix two other binding anomalies on the Order list.