VIX-2731 Fix FastPixel so it can be used in the effect.
FastPixel in most usages gets a full alpha bitmap passed into it. But in this case we want to let it create it, however the SetPixelAlpha method assumes the alpha is already set to max byte value. I added logic to assert that assumption by setting it to byte max value which will have no affect on other uses that pass in a bitmap with that to begin with.

Utilize the updated FastPixel to transfer the values from the particles to a render buffer to optimize the speed.

Fix reference to FastPixel to not copy local. This corrects an issue with how it is loaded.