VIX-2777 Improvements to inventory management.
Remove the hard coded list of vendors and their urls and instead retrieve it from the remote vixen web server. This is in a json file so added Newtonsoft.Json.12.0.2.nupkg and updated the other uses in the solution to use the latest also.

Centralize some of the http client logic into services.

Add logic to first look for a similarly named Vixen prop before importing the xModel.