VIX-2777 Improvements to support multiple models per Product.
Some products can have multiple models and the vendor already supports this. Added logic to handle the multiples and give the user the choice to pick which one they want to load.

Add support for multiple categories per product. The current vendor is not utilizing it, but the provisions for it are there.

Add basic model support for multiple images. The vendor is not utilizing that currently and the logic is not here yet to display them, but we can at least parse them from the inventory xml.

Added logic and a place holder for the pixel count. We will call it more generically Node Count.

Move and rename the Import xModel to a Tools menu as Vendor Browser.