VIX-391 VIX-679 Initial commit of work around opening unknown sequences.
A lot going on here. First add the logic to allow for the Element nodes to derive from an interface. I ported the work over from the virtual elements ticket to use here. This will allow the mapping of unknown effect/elements to a temporary placeholder element. This required a fair amount of refactoring to make all the places that used ElementNode use the interface.

Add logic to map the effects/elements that are unknown to a new ProxyElementNode class that will support enough info to manage them.

Add logic to the startup of the editor to recognize these elements/effect and prompt the user if they would like to map them to an existing element. IF so a mapping screen will appear to facilitate the process. Added the initial work around the new mapping screen and the logic to drive it.

Effects can now be mapped, but more work needs to be done to save the mapping for recall should the user need to do more that one.