VIX-391 VIX-679 More changes to get persistence fully working.
Upgrade Catel from 5.9 to 5.11 to fully support persisting as JSON. Upgraded it across the platform so we are on the current release. This included a point release to NewtonSoft JSON library from 12.0.1 to 12.0.2. There was a method name change in the new version which corresponded to a change in the CustomPropEditor that used that method. Appears to just be a name change.

Fixed up some issues and made the modified indicator work correctly. Added logic to update the command states when maps are saved. Overall it is much more usable now.

TODO Finish the logic that verifies a loaded map has all the source elements from the incoming sequence.