VIX-391 VIX-679 Finish the logic to update any opened map with missing sources.
Add logic to check newly opened maps to see if they have all the sources in them. If not ask the user if they want to add them.

Utilize the IOC to find MessageBoxService. Need to retrofit this in other areas as we incorporate Catel deeper into the WPF parts.

Remove the old view code that will not be used.