VIX-391 VIX-679 Refactor the MessageBoxService out to the common area.
This was starting to get duplicated, so I refactored it out to the common area and registered it with the IOC in the main VixenApplication so it could be used elsewhere. Refactored Custom Prop Editor to use it. Refactored the location in the Preview. The preview is not directly in a model class that can use the IOC, but it can still use the class from the new location.