VIX-2781 Fix hard crash when loading dependencies.
The System.Windows.Interactivity was being copied local in the effect editor and that caused a localization conflict between two loaded versions. This corrects the copy local issue and solves the crash.

I also found a few items in the Custom Prop Editor that were copying local when they should not be. I don;t think these were causing any immediate issues, but I cleaned them up to be sure.