VIX-2783 Fix the refresh issue with imported props in the Preview element tree.
When the element tree was switched to use the one from the display setup, the refresh mechanism that was in the preview was no longer working. Wire up a new mechanism to manage that.

Set the initial directory path to be the profile if it is not already set for imports.