VIX-2789 Add support for Face Mapping in the Custom Prop Editor.
Added a tab to manage Element properties into the prop editor. Separated the general prop info from a Element info tab. This tab provides the ability to see more details on the Element. Added a Face Component property to Elements and gave the ability for the user to assign a face component to an element or most likely group.

Added the ability to map the new Face Component property on the model to the Face Property upon loading a prop into the preview. This sets up the mapping according to the model definition.

Added the ability to import the xModel face definitions and map them to the new Face Component when importing a model. Now if the model is setup correctly the user will have very little additional work to do. They can alter that mapping with the mechanism above in the prop editor to suit their needs if necessary.