VIX-2792 Fix an issue in the where the set level effect fails to work.
The Set Level was changed to use a static intent instead of interpolating, but it needs a minimal interpolator that returns a copy of the value so that i tcan be modified downstream. Create interpolators for the discrete and lighting value types that return a copy of the value as the point in time intent.