VIX-2791 Invoke the element template that is appropriate for the prop when one is not selected.
Add logic to all the shapes to allow them to reconfigure themselves to a new ElementNode. All complex shapes but the Multi-String now support this. The Multi-String has some jenky logic in it and it needs more work to make it play well.

Once a prop is drawn, invoke the template wizard that applies so the user can create the elements if nothing was selected to link to.