VIX-2805 Further optimize the output section to eliminate wasteful cycles.
Convert to commands in the color break down filter. Before we were converting to an intensity state that was then further down the line converting to a command. There is not real difference other than the container. This eliminates one step. The only down side is we are now limiting ourselves to 8 bit data in the color path. No big deal as nothing currently uses bigger values.

Make some changes in the dimming filter to handle commands if it is behind the breakdown curve or continue to process intents if it is front. This makes me even more want to make a dimming breakdown filter and further optimize the chain.