VIX-2802 Add the logic to enable he user to enable compression on FPP Export.
Add a check box to the single export and the wizard format stage to allow the user to enable exporting with compression enabled in the new format.

Add some members to the export writer interface to better support detecting what the export format supports. Falcon, compression. Create a base class that has a lot of boiler plate stuff in it to lighten the content of the implementing classes. This helps get rid of some magic string stuff that was questionable at best.