VIX-2693 Improve some aspects of the scheduler.
Rework the logic around the SequenceManager and the SequenceAction to be able to detect if a sequence changes while the show is running. If so reload the new version and retire the old one. This may have render impact on the running show, but I assume the user takes that knowledge when they want to change it.

Add logic to detect render errors in the SequenceAction and mark the item as can't execute so it will bypass it instead of throwing logs full of errors. IF the user fixes and saves as above, then it can run when it comes around if it renders properly.