VIX-2803 Make the installer behave better when reboots are required.
Move the .NET and C++ redist to after the Vixen code install. Add flags to suppress the restart. Evaluate the response code of those installs and prompt the user to reboot at the end. Change the .NET installer to use the web install so the proper language packs will be installed and it will be as light as possible.