VIX-2805 A few more general optimizations.
Intern the common string color names for the breakdown filters to save memory when the channel counts rise. Change them to constants.

Reduce the default size of the command list in the DataPolicy. That will rarely be used.

Clean up the code flow a bit in the e1.31 controller. If there are no universes defined then don't process the commands.