VIX-391 VIX-679 Add direct source and destination tree mapping when the source tree is available.
Change the view to side by side tree mapping when the source tree is available.

Save the source tree to the map file for convenience. Improve the serialization times of saving and loading the map file by using Newtonsoft JSON directly. The CATEL serializer is very slow.

Add view models for the source tree to enable the source to destination labels and color coding to work without cluttering up the model classes.

TODO refactor this whole thing out to its own module.