VIX-391 VIX-679 Add the sequence package export wizard.
Add the wizard steps to create a sequence package of the sequences a user selects for export. This packages up the sequence, any dependent files for the effects, and the audio file if there is one and the user chooses to have it included.

TODO create the Import wizard steps to import this goodness.