VIX-391 VIX-679 More work on the input selection stage of import.
Wire in the map screens to create a map if one does not exist. Enable the wizard to retrieve the saved map file name to work from. Enable the wizard to inject the element tree file from the package archive to enable tree mapping.

Fixed a bug in the map screen that would not update the target element on the initial drop because the source id was not loaded yet.