VIX-391 VIX-679 Fix up a few smaller issues and improvements.
Set the sequence migration to popup message boxes in center of screen. Users may run into those messages and they currently popup all over the place. Not specifically related to this change, but makes the experience crappy if it happens in the import.

Remove the migration versions of sequences if migrations occur during sequence import to keep the sequence folder cleaner. They are not usable anyway because they have not been remapped.

Only set the filename to empty in the sequence editor for updated sequences if they do not reside in the profile sequence directory. This will prevent overriding sequences from other profiles, but let ones in the profile easily save to the same place.

Fix backing up in the wizard to the sequence screen that it properly handles having the correct sequences checked.