VIX-2811 Refactor how the tools strip visibility settings are saved.
Create a new format for saving the tool strips visibility based on the tool strip that owns it. The new format is more efficient that the original way by using a dictionary to look them up vs inserting across a large list.

Separate the color, gradient, and curves into profile specific files because they are per profile. Save the rest of them into the main file so they can be shared across profiles.