VIX-2811 Significant refactor to the Curve, Gradient, and Color Toolbar functions.
Remove all the duplicate logic that was copied from the individual libraries to manage adding, deleting, import, exporting of libraries. Exposed the same functions as internal from the individual library forms.

Refactor the drag / drop logic to work correctly. It was not waiting for the user to move the mouse before assuming a drag / drop was to take place.

Other code clean up and simplification.