VIX-391 VIX-679 Bug fixes and cosmetic polish.
Change the grid splitter to be narrow 1px black as suggested by feedback. Correct the views to allow the splitter to work correctly in either view.

Fix some missing paths that should be detecting if the sequence might be foreign to try and prevent the user over writing one they are importing from elsewhere.

Fix a bug that would crash if the user edited a map before the package file was selected.