VIX-391 VIX-679 Fix a few more issues.
The selection in the import sequences was not properly updating the can move next by triggering a wizard stage changes on the selection of sequences.

Consolidate the sequence module data under a common parent folder in the zip to keep things more organized. Update the export and import to use a SequenceData as the top level path. This lines up with the other data bits having a nice clean folder structure in the package zip file.