VIX-2819 Improve the performance of the Preview Setup screen.
Remove the render timer that rendered all the time on a schedule. Add logic to all the places that change things that affect the display to call the render routine and update the screen. This reduces render cycles to those that are actually needed to respond to changes. This significantly reduces the amount of CPU used in this form. As we allowing more setup in this screen, users will be spending more time in it and the improvement will be much needed.