VIX-2820 Optimize the linking in the preview setup.
Restructure the logic that sorts the columns in the preview elements linking to only adjust the columns once the operation has been completed.

Refactor the drag logic to be more efficient in how it selects the hovered target. Fix the drag enter and drag over methods to work correctly and set the proper effect when the element can be dropped and when it can't. This should eliminate the bizarre popup when it is dropped in the wrong place.

Make an optimization in the Element tree sorting mechanism to allow it to sort faster when elements have the same parent. This can speed up element selection a fair amount. Further optimization could be done later with more extensive refactoring.