VIX-2822 Optimize the color space conversions.
The focus was on the HSV to Color and the Color to HSV. Move around some of the math operations so they they only execute if they have to. Simplify the math by setting some constant factors that were being calculated all the time.

Make the return types what we need instead of running it through another conversion.