VIX-2811 Part two on attempts to fix ghost tool buttons.
I suspect the issue was coming with the save portion that was trying to gather all the state of the toolbars when closing. There was logic that set the whole toolbar visible if it was not to try and get its child items visible state. My theory is that causes layout inconsistencies and may lead to false values for visible. We are already updating the state on the fly when it changes except for the effects. I added the effects to that live update and then the state gathering at close is no longer required. This allows us to just save it. Less chance for issues to occur.

Cleaned up the PopulateToolStripStates method to only handle the initial population. This is probably not even necessary because the states will be updated as they are changed and if the state is not present it defaults to visible which is what we want.