VIX-837 Fix jittery movement.
The primary cause here was the Width value was being calculated incorrectly and as the objects moved, certain locations would cause values that would not let the Layout method execute to position the pixels in the proper place.

Clean up some other inefficiencies to make the overall layout math slightly faster.