VIX-2840 Virtualize the link elements tree loading in the Preview.
The ticket was focused on the speed of selection. There is not any logic around the treeview selection, so the speed is purely the cost of the control itself managing the selection. Virtualizing the tree makes it smaller, so it will load faster and maybe the selection logic will work better with a smaller foot print.

Fix a double cast in the expand logic in the setup element tree that was found when reusing the logic for this fix. Marginal improvement, but proper structure.