VIX-2876 Allow pixel specific effects to set their frame rate to match the default interval.
Added logic to the base class to obtain the frame rate from the default interval.

Added logic to the export wizard to set the default rate to match the export rate so the sequences will be rendered at the correct frame rate where pixel effects are involved.

Does not address exporting from the sequence editor where the loaded sequence is already rendered at the default rate.