VIX-2792 Part two of this fix to address additional issues.
When a Set Level is used on a group of discrete elements and a dimming curve is in place, those multiple elements will have a multiplier of the dimming curve for each child under the main group. The fix is to ensure each child element gets it's own intent vs trying to share a common one across them.