VIX-2889 Make adjustments in several effects to compensate for the speed variations when frame rate changes.
Changed Balls, Bars, Butterfly, Circles, ColorWash, Garlands, Life, Meteors, PinWheel, Shapes, Snowflakes, Spiral, SnowStorm around the speed timing calculation that is based on the frame interval that was fixed based on the assumed frame rate never changing. It is now a factor of the frame rate in relation to the assumed 20 FPS or 50 ms update.

Countdown had changes around the FPS math used to countdown the number of seconds along with the font size grow and shrink.