VIX-2890 Migrate all the package.config files into the projects as a PackageReference.
PackageReference is the new standard going forward and is necessary to migrate to the new project style for .NET Core support. Copy Local references in the old style are now handled with a <ExcludeAssets>runtime</ExcludeAssets> tag on the asset in the project file. Unfortunately this has to be manually added and the properties no longer show whether it is properly set. This will require more diligence on the developer.