VIX-759 Add the logic to make the speed setting work.
Introduced the SoundTouch library that allows for shifting the speed of the playback. It can do rate change or tempo change.

Wrapped the SoundTouch library in a sample provider based on one in the NAudio sample code to manage running the samples through it. It needs to buffer some and that offsets the position that we use for timing. Set the buffer at 20 ms and added logic that if the speed was normal that it bypasses and processes more real time.