VIX-2981 Fix the logic in the spin to avoid invalid values from breaking the effect.
In some cases depending on the settings and the element that the effect is placed on it can cause the effect to have render errors. This will try to detect the invalid value condition that can occur and use use values that will not cause an error. In all likelihood the effect may not do anything meaning full and the user will correct it. USer provided example was setting the Fixed Pulse interval to a value that combined was longer than the effect duration.