VIX-2890 More cleanup and version updates.
Remove old version files that are now covered in the default team city property files.
Remove the assembly updater as that is now done in a CI build process with dotnet-setversion.
Add the github-release exe that will create the Github release in the build process.
Update the CreateRelease.ps1 script to use the Github release program from the build folders.
Add the dotnet-setversion tool to update the assembly versions at build time.