VIX-2904 Add option to toggle the effect info tool tip.
Add menu item under the View to select to show the tool tip.
Add shortcut Ctrl+I that will toggle the tool tip and remove it from the drag box filter selection which is unlikely to be used.
Add logic to save the setting to the normal editor settings.
Add logic to the grid drawing routine to draw based on the setting.