VIX-2120 Refactor a few items to standardize and solve some problems.
Refactor new project files into SDK style projects to match the rest of the solution. This keeps them in line with the future of .NET. This change cascaded into some dependent projects to ensure that libraries are not copied local. Added proper Catel version as NuGet dependencies.

Make a change to the overall Theme for ItemsControl to remove the specific background color as it removes transparency that may be needed in some cases. The underlying controls should drive the background color.

Refactored the MainWindow to match the naming convention of Catel. It is now PolygonEditorView which aligns with the PolygonEditorViewModel. This allows us to use the Catel Window object as the base of the window which has other pattern benefits.

Refactor the Ok/Cancel buttons in the Polygon editor to use a WrapPanel instead of a UniformGrid so the styling will match the rest of the windows and use the Row Button Styles in the parent theme. Added background color attributes to the DockPanel and ToolBarTray to correct their background. It may be possible to move these up to the parent theme. The commands on the buttons need to be wired up to the model.