VIX-2601 Add exception handling onto the curve mouse handler.
The root of this tickets issue is not with the pinwheel itself, but with a generic error that happens in the mouse move handler in the curve editor. This appears in other effects and seems to manifest with the right series of events that cause some object to be null in the method. I suspect this is the curve itself being replaced but the log traces so far have not be specific enough. This adds logic to trap and eat the exception if it occurs to prevent the app from crashing. This method is called numerous times in a editing fashion and will likely recover just fine if it does not execute completely.