VIX-2750 Add progress reporting on startup.
Add a text enabled progress bar to the main admin screen to conserve space.
Refactor the profile loading logic to async and also to report progress. This has a bit more knowledge built into it about the total progress than I would like, but it serves the purpose for now.
Added a new progress bar to the common controls that allows text to be overlaid in the bar itself to save space.