VIX-2916 Implement GitHub Actions workflow for release builds
Implements workflow to compile the Vixen application, build installer
packages, and create a new release.

CreateReleaseNotes script updated to also output the release notes for
this version, transformed to markdown, for use by the Actions workflow.

Development builds and a pre-release are generated on a push to master.

Release builds and a normal release are generated on a new tag matching
a release version pattern (e.g. 1.2u3).