VIX-2920 Fix up Prop Editor fields.
The upgrade to newer versions of Catel exposed a few new properties in view models that the Property Grid was dynamically picking up. Added logic to override and add the browsable attribute to them to suppress them.

Added some tool tip descriptions to a few other fields to help the user with their meaning.

Populate the created by field with the system user name so it has some default.