VIX-2927 Improve Face imports for analog props and better effect defaults.
Add some logic to better set sensible defaults for color mapping on discrete props imported with face mapping. Ensure all elements with face mapping have a color assigned that matches the underlying elements. This should improve it just working out of the box. User can adjust colors as needed afterwards.

The Lipsync defaults to using phonemes instead of Mark Collections. But if we have a MArk collection with Phonemes, the user most likely wants to use the better workflow. So make that the default along with setting the outline on and the eyes open.