VIX-2750 Fix threading issues with the menu items.
Menu items are now being loaded in background threads, so guards were added to ensure the menus are added on the UI thread from the modules when they are creating app commands. This was isolated away from the modules in the AppCommand class. It does rely on the main form creating the first AppCommand to initialize it with the UI control.