VIX-2929 Improve how backups are kept when the user saves changes.
Prior to this the backups were made by copying the file to a new version. This could be corrupted also, so this will just attempt a move which should be a simple rename at the OS level and hopefully simpler. The backups will be named with a time stamp and purged once there are more than 3 to maintain the 3 backups. If corruption is detected, the backups will be renamed to the protected names so they do not get overwritten.