VIX-2933 Add ability to duplicate a preview.
Add UI elements and logic in the ConfigPreviews to duplicate selected previews.

Core changes to the make sure the clone methods on all the shapes work correctly and make a unlinked duplicate of the shapes. This looks more invasive than it actually was. All the shapes were touched as many of them did not have a proper clone override. Some of them did, but it was not making a proper deep clone of all items.