VIX-2934 Fix an issue where the first frame text would be black.
The level value was not being initialized before the initial render operation happened on the first frame. This was causing the first frame to have transparent text on the the initial render. It is a class level variable, so the next time it rendered it would have the last value which if positive would produce a non transparent frame. I moved the initialization of that variable to ensure it happens correctly.