VIX-2943 Enhance the Shockwave radius curve to use a scaled range based on the prop size.
Add logic to select the new scaled method and make it the default for new effects. Old effects will remain unscaled to allow clean backward compatibility.

Change the default curve for the radius to provide a better default with the scaling option.

Fix the acceleration rate curve default to act like the original did before the curve introduction.