VIX-2120
Added logic to select the line after converting a polygon or an ellipse to a line.
Fixed the SetupRender logic for the times of the pattern wipe.  At some point I got confused on the expected behavior.  When a stagger is involved the shapes should not complete at the same time.